pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "testfastapi"
version = "0.1.0"
description = "FastAPI测试项目"
authors = [
    {name = "graythorn", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.23.2",
    "sqlalchemy==2.0.23",
    "psycopg2==2.9.10",
    "python-dotenv==1.0.0",
    "alembic==1.12.1",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "pydantic[email]"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.uvicorn]
app = "app.main:app"
host = "0.0.0.0"
port = 8000
reload = true